feat: add command to cache rumble thumbnails
