+++ core/templatetags/thumbnails.py
@@ -1,4 +1,6 @@
 from django import template
+from django.conf import settings
+
 from ..utils.thumbnails import svg_placeholder as _svg_placeholder
 
 register = template.Library()
@@ -12,3 +14,15 @@ def svg_placeholder(label, alt=None):
     """
     src, alt_text = _svg_placeholder(label, alt)
     return {"src": src, "alt": alt_text}
+
+
+@register.filter
+def local_thumb(url: str) -> str:
+    """Return ``url`` if it is under ``MEDIA_URL``.
+
+    Any URL not starting with ``MEDIA_URL`` is treated as missing so templates
+    can fall back to placeholders instead of rendering remote thumbnails.
+    """
+    if url and url.startswith(settings.MEDIA_URL):
+        return url
+    return ""
