+++ core/tests/test_links.py
@@ -4,6 +4,7 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.template.loader import render_to_string
 from django.test import RequestFactory, TestCase
 from django.urls import reverse
+from django.conf import settings
 from io import BytesIO
 from PIL import Image
 
@@ -105,7 +106,7 @@ class PostLinkTests(TestCase):
             title="Link",
             content_url="https://example.com/article",
         )
-        post.thumbnail_url = "https://example.com/thumb.jpg"
+        post.thumbnail_url = f"{settings.MEDIA_URL}thumb.jpg"
         request = self.factory.get("/")
         html = render_to_string("partials/feed_card.html", {"post": post}, request=request)
         detail_url = post.get_absolute_url()
@@ -125,7 +126,7 @@ class PostLinkTests(TestCase):
             title="Link",
             content_url="https://example.com/article",
         )
-        post.thumbnail_url = "https://example.com/thumb.jpg"
+        post.thumbnail_url = f"{settings.MEDIA_URL}thumb.jpg"
         request = self.factory.get("/")
         html = render_to_string("core/partials/post_row.html", {"post": post}, request=request)
         detail_url = post.get_absolute_url()
