+++ core/tests/tests_backfill_thumbs_command.py
@@ -171,41 +171,3 @@ def test_backfill_thumbs_uses_fallback_after_og(monkeypatch):
     )
 
 
-@pytest.mark.django_db
-def test_backfill_thumbs_caches_rumble_thumbnails(monkeypatch):
-    cache.clear()
-    User = get_user_model()
-    user = User.objects.create_user("alice", password="pw")
-    com = Community.objects.create(slug="t", name="Test", title="Test", created_by=user)
-    post = Post.objects.create(
-        community=com,
-        author=user,
-        post_type="link",
-        title="Link",
-        content_url="https://rumble.com/v1abc",
-        thumbnail_url="https://sp.rmbl.ws/s8/1/v1abc.jpg",
-    )
-
-    from core.management.commands import backfill_thumbs
-
-    def fake_cache(origin_url):
-        return "/media/thumbs/rumble/cache.jpg"
-
-    monkeypatch.setattr(backfill_thumbs, "cache_remote_image", fake_cache)
-    call_command("backfill_thumbs", days=365)
-    post.refresh_from_db()
-    assert post.thumbnail_url == "/media/thumbs/rumble/cache.jpg"
-
-    post.thumbnail_url = "https://sp.rmbl.ws/s8/1/v1abc.jpg"
-    post.save(update_fields=["thumbnail_url"])
-    calls: list[str] = []
-
-    def fake_cache2(origin_url):
-        calls.append(origin_url)
-        return "/media/thumbs/rumble/cache2.jpg"
-
-    monkeypatch.setattr(backfill_thumbs, "cache_remote_image", fake_cache2)
-    call_command("backfill_thumbs", days=365, dry_run=True)
-    post.refresh_from_db()
-    assert post.thumbnail_url == "https://sp.rmbl.ws/s8/1/v1abc.jpg"
-    assert calls == ["https://sp.rmbl.ws/s8/1/v1abc.jpg"]
